# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Process-Sync-Api",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/incidents": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list incidents",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "list incidents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "default 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "default id.asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma seperated list of network-ids, used to filter the result",
                        "name": "network_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "process-instance-id, used to filter the result",
                        "name": "process_instance_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Incident"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/incidents/{networkId}/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get incident",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "get incident",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "incident id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Incident"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete incident",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incidents"
                ],
                "summary": "delete incident",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "incident id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/metadata/{networkId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list deployment metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment",
                    "metadata"
                ],
                "summary": "list deployment metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "network id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "deployment id",
                        "name": "deployment_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "camunda deployment id",
                        "name": "camunda_deployment_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DeploymentMetadata"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "deploymentmodel.ConditionalEvent": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "qos": {
                    "type": "integer"
                },
                "script": {
                    "type": "string"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                },
                "value_variable": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "deploymentmodel.Diagram": {
            "type": "object",
            "properties": {
                "svg": {
                    "type": "string"
                },
                "xml_deployed": {
                    "type": "string"
                },
                "xml_raw": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.Element": {
            "type": "object",
            "properties": {
                "bpmn_id": {
                    "type": "string"
                },
                "conditional_event": {
                    "$ref": "#/definitions/deploymentmodel.ConditionalEvent"
                },
                "group": {
                    "type": "string"
                },
                "message_event": {
                    "$ref": "#/definitions/deploymentmodel.MessageEvent"
                },
                "name": {
                    "type": "string"
                },
                "notification": {
                    "$ref": "#/definitions/deploymentmodel.Notification"
                },
                "order": {
                    "type": "integer"
                },
                "task": {
                    "$ref": "#/definitions/deploymentmodel.Task"
                },
                "time_event": {
                    "$ref": "#/definitions/deploymentmodel.TimeEvent"
                }
            }
        },
        "deploymentmodel.FilterCriteria": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "type": "string"
                },
                "characteristic_id": {
                    "description": "TODO: remove?",
                    "type": "string"
                },
                "device_class_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.GenericEventSource": {
            "type": "object",
            "properties": {
                "filter_ids": {
                    "type": "string"
                },
                "filter_type": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.IncidentHandling": {
            "type": "object",
            "properties": {
                "notify": {
                    "type": "boolean"
                },
                "restart": {
                    "type": "boolean"
                }
            }
        },
        "deploymentmodel.MessageEvent": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "flow_id": {
                    "type": "string"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                },
                "use_marshaller": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.Notification": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.ProcessStartParameter": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "deploymentmodel.Selection": {
            "type": "object",
            "properties": {
                "filter_criteria": {
                    "$ref": "#/definitions/deploymentmodel.FilterCriteria"
                },
                "selected_device_group_id": {
                    "type": "string"
                },
                "selected_device_id": {
                    "type": "string"
                },
                "selected_generic_event_source": {
                    "$ref": "#/definitions/deploymentmodel.GenericEventSource"
                },
                "selected_import_id": {
                    "type": "string"
                },
                "selected_path": {
                    "$ref": "#/definitions/deviceselectionmodel.PathOption"
                },
                "selected_service_id": {
                    "type": "string"
                },
                "selection_options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.SelectionOption"
                    }
                }
            }
        },
...
        "deploymentmodel.Task": {
            "type": "object",
            "properties": {
                "parameter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "prefer_event": {
                    "type": "boolean"
                },
                "retries": {
                    "type": "integer"
                },
                "selection": {
                    "$ref": "#/definitions/deploymentmodel.Selection"
                }
            }
        },
        "deploymentmodel.TimeEvent": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "deviceselectionmodel.Configurable": {
            "type": "object",
            "properties": {
                "aspect_node": {
                    "$ref": "#/definitions/devicemodel.AspectNode"
                },
                "characteristic_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "deviceselectionmodel.PathOption": {
            "type": "object",
            "properties": {
                "aspectNode": {
                    "$ref": "#/definitions/devicemodel.AspectNode"
                },
                "characteristicId": {
                    "type": "string"
                },
                "configurables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deviceselectionmodel.Configurable"
                    }
                },
                "functionId": {
                    "type": "string"
                },
                "isVoid": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "github_com_SENERGY-Platform_process-sync_pkg_model_camundamodel.Variable": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {},
                "valueInfo": {}
            }
        },
...
        "model.DeploymentMetadata": {
            "type": "object",
            "properties": {
                "camunda_deployment_id": {
                    "type": "string"
                },
                "deployment_model": {
                    "$ref": "#/definitions/model.DeploymentWithEventDesc"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "network_id": {
                    "type": "string"
                },
                "process_parameter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_SENERGY-Platform_process-sync_pkg_model_camundamodel.Variable"
                    }
                },
                "sync_date": {
                    "type": "string"
                }
            }
        },
        "model.DeploymentWithEventDesc": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "device_id_to_local_id": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "diagram": {
                    "$ref": "#/definitions/deploymentmodel.Diagram"
                },
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.Element"
                    }
                },
                "event_descriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EventDesc"
                    }
                },
                "executable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "incident_handling": {
                    "$ref": "#/definitions/deploymentmodel.IncidentHandling"
                },
                "name": {
                    "type": "string"
                },
                "service_id_to_local_id": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "start_parameter": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/deploymentmodel.ProcessStartParameter"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "model.EventDesc": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "type": "string"
                },
                "characteristic_id": {
                    "description": "marshaller info",
                    "type": "string"
                },
                "deployment_id": {
                    "description": "search info",
                    "type": "string"
                },
                "device_group_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "import_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "qos": {
                    "type": "integer"
                },
                "script": {
                    "description": "worker info",
                    "type": "string"
                },
                "service_for_marshaller": {
                    "description": "set by event-manager\nmay be\n\t- the service from EventDesc.ConditionalEvent.Selection.SelectedServiceId\n\t- or an artificial service for EventDesc.ConditionalEvent.Selection.SelectedImportId",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Service"
                        }
                    ]
                },
                "service_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "value_variable": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.Incident": {
            "type": "object",
            "properties": {
                "deployment_name": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "external_task_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_placeholder": {
                    "type": "boolean"
                },
                "marked_as_missing": {
                    "type": "boolean"
                },
                "marked_for_delete": {
                    "type": "boolean"
                },
                "msg_version": {
                    "description": "from version 3 onward will be set in KafkaIncidentsCommand and be copied to this field",
                    "type": "integer"
                },
                "network_id": {
                    "type": "string"
                },
                "process_definition_id": {
                    "type": "string"
                },
                "process_instance_id": {
                    "type": "string"
                },
                "sync_date": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Content": {
            "type": "object",
            "properties": {
                "content_variable": {
                    "$ref": "#/definitions/models.ContentVariable"
                },
                "id": {
                    "type": "string"
                },
                "protocol_segment_id": {
                    "type": "string"
                },
                "serialization": {
                    "$ref": "#/definitions/models.Serialization"
                }
            }
        },
...
        "models.Interaction": {
            "type": "string",
            "enum": [
                "event",
                "request",
                "event+request"
            ],
            "x-enum-varnames": [
                "EVENT",
                "REQUEST",
                "EVENT_AND_REQUEST"
            ]
        },
        "models.Serialization": {
            "type": "string",
            "enum": [
                "xml",
                "json",
                "plain-text"
            ],
            "x-enum-varnames": [
                "XML",
                "JSON",
                "PlainText"
            ]
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attribute"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "interaction": {
                    "$ref": "#/definitions/models.Interaction"
                },
                "local_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "protocol_id": {
                    "type": "string"
                },
                "service_group_key": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}